Minor tweaks.

+++ src/solver.jl
@@ -164,7 +164,7 @@ end
 
 function trial(solver::DESPOTSolver, pomdp::POMDP, node::VNode, n_trials::Int64)
 
-    n_nodes_added::Int64 = 0
+    n_nodes_added = 0
     ub::Float64 = 0.0
     
     if (node.depth >= solver.config.search_depth) || isterminal(pomdp, node.particles[1].state)
